Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 // Function to perform the calculation
-double calculate(double num1, double num2, char operator) {
-    double result;
+int calculate(int num1, int num2, char operator) {
+    int result;
     switch (operator) {
         case '+':
             result = num1 + num2;
@@ -30,20 +30,20 @@ double calculate(double num1, double num2, char operator) {
 }
 
 int main() {
-    double num1, num2;
+    int num1, num2;
     char operator;
 
     // First calculation: 5 + 3
     num1 = 5;
     operator = '+';
     num2 = 3;
-    printf("Result of %.2lf %c %.2lf: %.2lf\n", num1, operator, num2, calculate(num1, num2, operator)); // Output: 8
+    printf("Result of %d %c %d: %d\n", num1, operator, num2, calculate(num1, num2, operator)); // Output: 8
 
     // Second calculation: 10 / 2
     num1 = 10;
     operator = '/';
     num2 = 2;
-    printf("Result of %.2lf %c %.2lf: %.2lf\n", num1, operator, num2, calculate(num1, num2, operator)); // Output: 5
+    printf("Result of %d %c %d: %d\n", num1, operator, num2, calculate(num1, num2, operator)); // Output: 5
 
     return 0;
 }
